unit tests from resource_assembler.rs moved to tests/unit/filters/resource_assembler.rs
